external decode service(s) for ubuntu
