+++ app/logging/README.txt
@@ -1,1 +1,5 @@
-Logging is still being worked on r.n and subject to constant change
+Models for Logging -
+
+HUDLogMessagesModel: Model for showing log messages in the HUD (3 seconds timeout)
+
+LogMessagesMode: Model for showing logs in its own UI (list-lke) element
