+++ app/telemetry/models/camerastreammodel.h
@@ -20,9 +20,10 @@ public:
     const int m_camera_index;
     static CameraStreamModel& instance(int cam_index);
 public:
-    // NOTE: hacky right now, since it is a param but we also want to display it in the HUD
-    L_RO_PROP(QString,curr_set_video_codec,set_curr_set_video_codec,"N/A")
+    // Width x height @ fps for a camera
     L_RO_PROP(QString,curr_set_video_format,set_curr_set_video_format,"")
+    // Width x height @ fps and codec for camera
+    L_RO_PROP(QString,curr_set_video_format_and_codec,set_curr_set_video_format_and_codec,"N/A")
     //
     // The following stats are kinda camera related, kinda wb link related
     // They area also only generated by OpenHD air unit.
@@ -44,6 +45,9 @@ public:
     L_RO_PROP(int,curr_fec_percentage,set_curr_curr_fec_percentage,-1)
     L_RO_PROP(int,curr_keyframe_interval,set_curr_curr_keyframe_interval,-1)
     L_RO_PROP(bool,air_recording_active,set_air_recording_active,false)
+    L_RO_PROP(QString,air_tx_packets_per_second_and_bits_per_second,set_air_tx_packets_per_second_and_bits_per_second,"N/A")
+    // UI might change slightly depending on the camera type
+    L_RO_PROP(int,camera_type,set_camera_type,-1)
     //
     // These are generated by the OpenHD ground unit
     //
@@ -54,11 +58,6 @@ public:
     L_RO_PROP(qint64,video0_count_fragments_recovered,set_video0_count_fragments_recovered,-1)
     L_RO_PROP(QString,curr_video0_fec_decode_time_avg_min_max,set_curr_video0_fec_decode_time_avg_min_max,"avg na, min na, max na")
     //
-    L_RO_PROP(QString,air_tx_packets_per_second_and_bits_per_second,set_air_tx_packets_per_second_and_bits_per_second,"N/A")
-    //L_RO_PROP(QString,gnd_rx_packets_per_second_and_bits_per_second,set_gnd_rx_packets_per_second_and_bits_per_second,"N/A")
-    L_RO_PROP(QString,lulu_curr_video_codec_and_format,set_lulu_curr_video_codec_and_format,"N/A")
-    // UI might change slightly depending on the camera type
-    L_RO_PROP(int,camera_type,set_camera_type,-1)
 public:
     // generated by wb / link
     void update_mavlink_openhd_stats_wb_video_air(const mavlink_openhd_stats_wb_video_air_t &msg);
@@ -73,6 +72,26 @@ public:
 private:
     // Do not completely pollute the Log messages model
     int m_n_mismatch_has_been_logged=0;
+    // FEAUTURE - Re-start decoding if resolution / framerate of a camera has changed
+    struct ResolutionFramerate{
+        int width=-1;
+        int height=-1;
+        int framerate=-1;
+        bool operator==(const ResolutionFramerate &o) const {
+            return this->framerate==o.framerate && this->width==o.width && this->height==o.height;
+        }
+        bool operator !=(const ResolutionFramerate &o) const {
+            return !(*this==o);
+        }
+        bool is_any_invalid()const{
+            return (width==-1 || width==0 ||
+                   height==-1 || height==0 ||
+                    framerate==-1 || framerate==0);
+        }
+    };
+    static std::string resolution_framerate_to_string(const ResolutionFramerate& data);
+    ResolutionFramerate m_curr_res_framerate{};
+
 };
 
 #endif // AIRCAMERAMODEL_H
