+++ systemd/x86_ubuntu_h265_decode.service
@@ -1,0 +1,15 @@
+[Unit]
+Description=ubuntu x86 h264 decode service
+
+[Service]
+User=root
+
+# Video decode via mpp, started by QOpenHD if needed (and stopped if needed)
+ExecStart=/bin/sh -c "gst-launch-1.0 -v udpsrc port=5600 caps = "application/x-rtp, media=(string)video, encoding-name=(string)H265" ! rtph265depay ! decodebin ! videoconvert ! autovideosink"
+
+Restart=always
+RestartSec=2
+
+[Install]
+WantedBy=multi-user.target
+
